+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added space production report
 - added number of occupants to space
 - added space cascader to energy flow diagram in myems-web
-- 
+- added tbl_spaces_distribution_systems to system database
 ### Changed
 - changed base docker image from python:silm to python:3.10-slim
 - udpated demo database scripts